// packages/react/src/lib/components/primary-navigation-utility-menu/primary-navigation-utility-menu.d.ts
import * as React from 'react';
export declare const PdsPrimaryNavigationUtilityMenu: React.FunctionComponent<Partial<{
    variant: "inverted" | "default";
}> & {
    ref?: React.Ref<any> | undefined;
} & {
    style?: React.CSSProperties | undefined;
} & Partial<{}> & {
    id?: string | undefined;
    slot?: string | undefined;
    className?: string | undefined;
    children?: React.ReactNode;
}>;